pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tjf"
version = "0.0.1"
description = ""
authors = ["Taavi Väänänen <hi@taavi.wtf>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
requests = "^2.31.0"
toolforge-weld = "^1.6.11"
pydantic = "^2.6.4"
gevent = "^24.2.1"
fastapi = "^0.116.1"
prometheus-fastapi-instrumentator = "^7.1.0"
uvicorn = "^0.35.0"

[tool.poetry.group.dev.dependencies]
openapi-schema-validator = "^0.6.2"
types-requests = "^2.32.0.20250306"
types-pyyaml = "^6.0.12.20250326"
types-setuptools = "^76.0.0.20250313"
prometheus-client = "^0.21.1"
pytest = "^8.3.5"
requests-mock = "^1.12.1"
typing-extensions = "^4.14.0"

[tool.black]
line-length = 99
target-version = ['py311']

[tool.isort]
profile = "black"
skip_gitignore = true
